+++ src/__tests__/ViewabilityHelper.test.ts
@@ -2,7 +2,7 @@ import { Dimension, Layout } from "recyclerlistview";
 
 import CustomError from "../errors/CustomError";
 import ExceptionList from "../errors/ExceptionList";
-import ViewabilityHelper from "../ViewabilityHelper";
+import ViewabilityHelper from "../viewability/ViewabilityHelper";
 
 describe("ViewabilityHelper", () => {
   const viewableIndicesChanged = jest.fn();
