+++ src/native/auto-layout/AutoLayoutViewNativeComponent.web.ts
@@ -1,0 +1,8 @@
+import React from "react";
+import { View } from "react-native";
+
+import { AutoLayoutViewNativeComponentProps } from "./AutoLayoutViewNativeComponentProps";
+
+const AutoLayoutViewNativeComponent =
+  View as any as React.Component<AutoLayoutViewNativeComponentProps>;
+export default AutoLayoutViewNativeComponent;
