+++ src/native/auto-layout/AutoLayoutViewNativeComponentProps.ts
@@ -1,0 +1,13 @@
+export interface OnBlankAreaEvent {
+  nativeEvent: {
+    offsetStart: number;
+    offsetEnd: number;
+  };
+}
+
+type OnBlankAreaEventHandler = (event: OnBlankAreaEvent) => void;
+
+export interface AutoLayoutViewNativeComponentProps {
+  onBlankAreaEvent: OnBlankAreaEventHandler;
+  enableInstrumentation: boolean;
+}
