+++ src/native/cell-container/CellContainer.web.ts
@@ -1,0 +1,6 @@
+import { View } from "react-native";
+/**
+ * On web we use a view instead of cell container till we build native web implementations
+ */
+const CellContainer = View;
+export default CellContainer;
